requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
attrs==21.2.0
blinker==1.4
click==8.0.1
confluent-kafka==1.7.0
dacite==1.6.0
elastic-apm==6.3.3
Flask==2.0.1
Flask-Cors==3.0.10
gunicorn==20.1.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.12.2
marshmallow-enum==1.5.1
marshmallow-union==0.1.15.post1
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pymongo==3.11.4
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.18.0
python-keycloak==0.26.1
pytz==2021.1
regex==2021.7.6
six==1.16.0
text-unidecode==1.3
toml==0.10.2
ujson==4.0.2
Werkzeug==2.0.1
